requirements.txt
autopep8==2.0.0
click==8.1.3
Flask==2.2.2
Flask-Cors==3.0.10
geojson==2.5.0
importlib-metadata==5.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
psycopg2-binary==2.9.5
pycodestyle==2.10.0
Shapely==1.8.5.post1
six==1.16.0
tomli==2.0.1
Werkzeug==2.2.2
zipp==3.11.0
